+++ package.json
@@ -6,6 +6,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@supabase/supabase-js": "^2.45.4",
     "bootstrap": "^5.3.3",
     "next": "^14.2.9",
     "react": "^18.3.1",
